Album Cover Catch
Activates and deactivates a transform effect on a "Song Title" and "Artist Name" layer depending on the availability of an album cover.

Bind this script to the visibility of a placeholder image and add a Transform effect called "Song Transform" with Vertex mode for this to work. You'll also need to change the names of your layers or change the names listed here. I've soft-coded them to make it easy.